+++ backasync/src/main/java/com/sous/backasync/launch/interfaces/ModuleInsertingBackAsyncInterface.java
@@ -13,5 +13,5 @@ public interface ModuleInsertingBackAsyncInterface {
 
     Integer getModuleInsert(@NonNull Bundle bundleModuleBack);
     // TODO: 28.01.2025
-    Integer getModuleInsert(@NonNull String Таблица,@NonNull ContentValues contentValuesModuleBackAsync);
+    Long getModuleInsert(@NonNull String Таблица, @NonNull ContentValues contentValuesModuleBackAsync);
 }
